Restructure the Shared folder structure
