+++ Sources/Shared/Classes/ListAdapter.swift
