+++ Sources/Shared/Classes/SpotFactory.swift
