+++ Sources/Shared/Enums/SpotsAnimation.swift
