+++ Sources/Shared/Extensions/SpotsProtocol+LiveEditing.swift
