+++ Sources/Shared/Protocols/Listable.swift
