+++ Sources/Shared/Protocols/SpotAdapter.swift
