+++ Sources/Shared/Protocols/SpotConfigurable.swift
