+++ Sources/Shared/Protocols/SpotsDelegates.swift
