+++ Sources/Shared/Structs/Component.swift
