+++ Sources/Shared/Structs/GridableMeta.swift
@@ -1,0 +1,8 @@
+public struct GridableMeta {
+  public struct Key {
+    public static let sectionInsetTop = "inset-top"
+    public static let sectionInsetLeft = "inset-left"
+    public static let sectionInsetRight = "inset-right"
+    public static let sectionInsetBottom = "inset-bottom"
+  }
+}
