+++ Sources/Shared/Structs/SpotCache.swift
